requirements.txt
setuptools~=60.5.0
pandas~=1.3.3
selenium~=3.141.0
pymongo~=3.12.0
mysql-connector-python~=8.0.25
paramiko~=2.8.0
deathbycaptcha~=0.1.1
fake_useragent
lxml
beautifulsoup4~=4.10.0
2captcha-python==1.0.3
capmonster_python
python-anticaptcha==0.6.0
requests~=2.22.0
numexpr
psycopg2~=2.9.3
Flask~=2.1.2
TwoCaptcha~=0.0.1

psutil
gitPython